Create CharacterSelect(Not Funtional)
It was only a idea, but I couldn't find a way to complete it.  I spent not a lot of time on it, but if I had a bit of patience, I could have made it a reality. The way the game turned out also didn't need this, so it would have been pointless.